backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
click==8.3.0
coverage==7.11.3
dj-rest-auth==7.0.1
Django==5.2.8
django-appconf==1.2.0
django-cors-headers==4.9.0
django-imagekit==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
Faker==38.0.0
h11==0.16.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pilkit==3.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-sugar==1.1.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
termcolor==3.2.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.38.0
